D&Dで異常なデータ登録がなくなるように修正 (#99)
Issue: #95 

* refactor: EventEntryの論理削除を行っている実装を1つにまとめる

`if (minrEvent.deleted) {}` の処理を追加しているが、
論理削除されたデータを再度論理削除したとしても何も変わらないので、
念のための修正である。

* refactor: イベント保存時にバリデーションを行うように対応

drag&dropの対策を行って、再現しなくなったが、どういう状況にしても、
'no title' というイベントが作成されたことは事実で、サービス側に
必須チェックのバリデーションがないということも問題の1つではあるので、
その実装を行った。

* fix: #95 D&Dでイベントが変わったときの座標計算を修正

もともとD&Dでイベントが変更されたときの再計算は実装されていて、
そのあと同時間帯に複数重なるケースの実装を行ったが、X方向の再計算が
抜けていたために、重なっていない状態から重なる状態にD&Dされたときに、
Xの座標が更新されなくて、イベントが重なって表示されてしまっていた。
EntryTimeCellが変わったら、x、yの両方を再計算するように修正。

+++ src/main/services/CalendarSyncProcessorImpl.ts
@@ -180,6 +180,9 @@ export class CalendarSyncProcessorImpl implements ITaskProcessor {
     }
     for (const minrEvent of minrEventsMap.values()) {
       if (minrEvent.externalEventEntryId) {
+        if (minrEvent.deleted) {
+          continue;
+        }
         await this.deleteMinrEvent(minrEvent);
         updateCount++;
       } else if (!minrEvent.externalEventEntryId) {
@@ -203,13 +206,13 @@ export class CalendarSyncProcessorImpl implements ITaskProcessor {
   }
 
   private async updateMinrEvent(minr: EventEntry, external: ExternalEventEntry): Promise<void> {
-    console.log('updateMinrEvent', minr, external);
+    console.log('updateMinrEvent', minr.id, minr, external);
     EventEntryFactory.updateFromExternal(minr, external);
     await this.eventEntryService.save(minr);
   }
 
   private async deleteMinrEvent(minr: EventEntry): Promise<void> {
-    console.log('deleteMinrEvent', minr);
+    console.log('deleteMinrEvent', minr.id, minr);
     EventEntryFactory.updateLogicalDelete(minr);
     await this.eventEntryService.save(minr);
   }
@@ -218,7 +221,7 @@ export class CalendarSyncProcessorImpl implements ITaskProcessor {
     calendarSetting: CalendarSetting,
     minr: EventEntry
   ): Promise<void> {
-    console.log('newExternalEvent', minr);
+    console.log('newExternalEvent', minr.id, minr);
     const external = ExternalEventEntryFactory.createFromMinr(minr, calendarSetting.calendarId);
     const updated = await this.externalCalendarService.saveEvent(external);
     EventEntryFactory.updateFromExternal(minr, updated);
@@ -226,7 +229,7 @@ export class CalendarSyncProcessorImpl implements ITaskProcessor {
   }
 
   private async updateExternalEvent(external: ExternalEventEntry, minr: EventEntry): Promise<void> {
-    console.log('updateExternalEvent', external, minr);
+    console.log('updateExternalEvent', minr.id, external, minr);
     ExternalEventEntryFactory.updateFromMinr(external, minr);
     const updated = await this.externalCalendarService.saveEvent(external);
     EventEntryFactory.updateFromExternal(minr, updated);
